agent_garden/requirements.txt
flask>=3.0.0
gunicorn>=21.2.0
google-genai>=1.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
celery>=5.3.0
redis>=5.0.0
pymysql>=1.1.0
